# evenStore: Optimize rawKV encoding and decoding, compressing