--- a/logservice/eventstore/eventstore.go
+++ b/logservice/eventstore/eventstore.go
@@ -18,6 +18,7 @@ import (
 	"github.com/flowbehappy/tigate/pkg/metrics"
 	"github.com/flowbehappy/tigate/pkg/mounter"
 	"github.com/flowbehappy/tigate/utils"
+	"github.com/klauspost/compress/zstd"
 	"github.com/pingcap/log"
 	"github.com/pingcap/tidb/pkg/kv"
 	"github.com/pingcap/tiflow/pkg/pdutil"
@@ -123,6 +124,9 @@ type eventStore struct {
 		dispatcherMap   *utils.BtreeMap[*heartbeatpb.TableSpan, *spanState]
 		subscriptionMap map[logpuller.SubscriptionID]*spanState
 	}
+
+	encoder *zstd.Encoder
+	decoder *zstd.Decoder
 }
 
 const dataDir = "event_store"
@@ -160,6 +164,16 @@ func NewEventStore(
 	if err != nil {
 		log.Panic("fail to remove path")
 	}
+	// Create the zstd encoder
+	encoder, err := zstd.NewWriter(nil)
+	if err != nil {
+		log.Panic("Failed to create zstd encoder", zap.Error(err))
+	}
+
+	decoder, err := zstd.NewReader(nil)
+	if err != nil {
+		log.Panic("Failed to create zstd decoder", zap.Error(err))
+	}
 
 	store := &eventStore{
 		pdClock:     pdClock,
@@ -168,6 +182,8 @@ func NewEventStore(
 		eventChs:    make([]chan eventWithSpanState, 0, dbCount),
 
 		gcManager: newGCManager(),
+		encoder:   encoder,
+		decoder:   decoder,
 	}
 	// TODO: update pebble options
 	// TODO: close pebble db at exit
@@ -346,6 +362,7 @@ func (e *eventStore) GetIterator(dispatcherID common.DispatcherID, dataRange *co
 		startTs:      dataRange.StartTs,
 		endTs:        dataRange.EndTs,
 		rowCount:     0,
+		decoder:      e.decoder,
 	}, nil
 }
 
@@ -439,7 +456,8 @@ func (e *eventStore) handleEvents(ctx context.Context, db *pebble.DB, inputCh <-
 		}
 		key := EncodeKey(uint64(item.state.span.TableID), item.raw)
 		value := item.raw.Encode()
-		if err := batch.Set(key, value, pebble.NoSync); err != nil {
+		compressedValue := e.encoder.EncodeAll(value, nil)
+		if err := batch.Set(key, compressedValue, pebble.NoSync); err != nil {
 			log.Panic("failed to update pebble batch", zap.Error(err))
 		}
 	}
@@ -520,6 +538,7 @@ type eventStoreIter struct {
 	startTs  uint64
 	endTs    uint64
 	rowCount int64
+	decoder  *zstd.Decoder
 }
 
 func (iter *eventStoreIter) Next() (*common.RawKVEntry, bool, error) {
@@ -533,9 +552,13 @@ func (iter *eventStoreIter) Next() (*common.RawKVEntry, bool, error) {
 
 	key := iter.innerIter.Key()
 	value := iter.innerIter.Value()
+	decompressedValue, err := iter.decoder.DecodeAll(value, nil)
+	if err != nil {
+		log.Panic("failed to decompress value", zap.Error(err))
+	}
 	_, startTS, commitTS := DecodeKey(key)
 	rawKV := &common.RawKVEntry{}
-	rawKV.Decode(value)
+	rawKV.Decode(decompressedValue)
 	isNewTxn := false
 	if iter.prevCommitTS == 0 || (startTS != iter.prevStartTS || commitTS != iter.prevCommitTS) {
 		isNewTxn = true