--- a/logservice/logpuller/region_change_event_processor.go
+++ b/logservice/logpuller/region_change_event_processor.go
@@ -21,6 +21,7 @@ import (
 
 	"github.com/flowbehappy/tigate/pkg/common"
 	"github.com/flowbehappy/tigate/pkg/metrics"
+	"github.com/klauspost/compress/zstd"
 	"github.com/pingcap/errors"
 	"github.com/pingcap/kvproto/pkg/cdcpb"
 	"github.com/pingcap/log"
@@ -63,16 +64,26 @@ type changeEventProcessor struct {
 
 	metricProcessDuration prometheus.Observer
 	metricTotalDuration   prometheus.Observer
+
+	zstdEncoder *zstd.Encoder
 }
 
 func newChangeEventProcessor(id uint, client *SubscriptionClient) *changeEventProcessor {
 	clientID := client.id.String()
 	workerID := strconv.Itoa(int(id))
+
+	// Create the zstd encoder
+	encoder, err := zstd.NewWriter(nil)
+	if err != nil {
+		log.Fatal("Failed to create zstd encoder", zap.Error(err))
+	}
+
 	return &changeEventProcessor{
 		client:                client,
 		inputCh:               make(chan statefulEvent, 64), // 64 is an arbitrary number.
 		metricProcessDuration: metrics.RegionWorkerProcessDuration.WithLabelValues(clientID, workerID),
 		metricTotalDuration:   metrics.RegionWorkerTotalDuration.WithLabelValues(clientID, workerID),
+		zstdEncoder:           encoder, // Initialize the encoder
 	}
 }
 
@@ -168,10 +179,10 @@ func (w *changeEventProcessor) handleEventEntry(ctx context.Context, x *cdcpb.Ev
 		zap.Uint64("subscriptionID", uint64(state.region.subscribedSpan.subID)),
 		zap.Int64("tableID", tableID),
 		zap.Int("rows", len(x.Entries.GetEntries())))
-	return handleEventEntry(x, startTs, state, emit, tableID)
+	return w.doHandle(x, startTs, state, emit, tableID)
 }
 
-func handleEventEntry(
+func (w *changeEventProcessor) doHandle(
 	x *cdcpb.Event_Entries_,
 	startTs uint64,
 	state *regionFeedState,
@@ -190,7 +201,7 @@ func handleEventEntry(
 				zap.Stringer("span", &state.region.span))
 
 			for _, cachedEvent := range state.matcher.matchCachedRow(true) {
-				revent, err := assembleRowEvent(regionID, cachedEvent)
+				revent, err := w.assembleRowEvent(regionID, cachedEvent)
 				if err != nil {
 					return errors.Trace(err)
 				}
@@ -210,7 +221,7 @@ func handleEventEntry(
 				return errUnreachable
 			}
 
-			revent, err := assembleRowEvent(regionID, entry)
+			revent, err := w.assembleRowEvent(regionID, entry)
 			if err != nil {
 				return errors.Trace(err)
 			}
@@ -249,7 +260,7 @@ func handleEventEntry(
 				return errUnreachable
 			}
 
-			revent, err := assembleRowEvent(regionID, entry)
+			revent, err := w.assembleRowEvent(regionID, entry)
 			if err != nil {
 				return errors.Trace(err)
 			}
@@ -267,7 +278,7 @@ func handleEventEntry(
 	return nil
 }
 
-func assembleRowEvent(regionID uint64, entry *cdcpb.Event_Row) (regionFeedEvent, error) {
+func (w *changeEventProcessor) assembleRowEvent(regionID uint64, entry *cdcpb.Event_Row) (regionFeedEvent, error) {
 	var opType common.OpType
 	switch entry.GetOpType() {
 	case cdcpb.Event_Row_DELETE:
@@ -278,20 +289,30 @@ func assembleRowEvent(regionID uint64, entry *cdcpb.Event_Row) (regionFeedEvent,
 		return regionFeedEvent{}, cerror.ErrUnknownKVEventType.GenWithStackByArgs(entry.GetOpType(), entry)
 	}
 
-	revent := regionFeedEvent{
-		RegionID: regionID,
-		Val: &common.RawKVEntry{
-			OpType:   opType,
-			Key:      entry.Key,
-			Value:    entry.GetValue(),
-			StartTs:  entry.StartTs,
-			CRTs:     entry.CommitTs,
-			RegionID: regionID,
-			OldValue: entry.GetOldValue(),
-		},
+	var compressedValue []byte
+	var compressOldValue []byte
+	// Use the pre-created encoder
+	if len(entry.GetValue()) != 0 {
+		compressedValue = w.zstdEncoder.EncodeAll(entry.GetValue(), nil)
+	}
+	if len(entry.GetOldValue()) != 0 {
+		compressOldValue = w.zstdEncoder.EncodeAll(entry.GetOldValue(), nil)
 	}
 
-	return revent, nil
+	return regionFeedEvent{
+		RegionID: regionID,
+		Val: &common.RawKVEntry{
+			OpType:       opType,
+			Key:          entry.Key,
+			Value:        compressedValue,
+			StartTs:      entry.StartTs,
+			CRTs:         entry.CommitTs,
+			RegionID:     regionID,
+			OldValue:     compressOldValue,
+			CompressType: common.CompressTypeZstd,
+		},
+	}, nil
+
 }
 
 func (w *changeEventProcessor) handleResolvedTs(ctx context.Context, batch resolvedTsBatch) {
@@ -342,5 +363,12 @@ func (w *changeEventProcessor) advanceTableSpan(ctx context.Context, batch resol
 				return
 			}
 		}
+	}
+}
+
+// Add a Close method to properly release resources
+func (w *changeEventProcessor) Close() {
+	if w.zstdEncoder != nil {
+		w.zstdEncoder.Close()
 	}
 }