<p>This PR reduces the cost of encoding and decoding RawKV and improves throughput.</p>

Row size | Master | This PR
-- | -- | --
4KB | 40MB/s | 110+ MB/s, lag 2-5s, CPU 400%
100KB | 40MB/s | 266 MB/s, lag 2-5s, CPU 550%


<!-- notionvc: 9fc789b4-30a9-463b-8090-c42a3e40acfa -->

Note that the results in the rightmost column do not represent the program's performance limit, but are constrained by the amount of traffic we can write upstream.